+++ src/Tivins/PhpCore/DotEnv.php
@@ -3,13 +3,19 @@ declare(strict_types=1);
 
 namespace Tivins\PhpCore;
 
+use Exception;
+
 class DotEnv
 {
     /**
      * Override the environment variables with the lines in the file.
+     * @throws Exception if the file does not exist
      */
     public static function loadFile(string $filename): void
     {
+        if (!file_exists($filename)) {
+            throw new Exception("File not found: $filename");
+        }
         $lines = file($filename);
         self::loadLines($lines);
     }